+++ packages/visionOS/web-spatial/JSBCommand.swift
@@ -272,6 +272,10 @@ struct UpdateSpatializedStatic3DElementProperties: SpatializedElementProperties 
     let modelTransform: [Double]?
     let autoplay: Bool?
     let loop: Bool?
+    let animationPaused: Bool?
+    let playbackRate: Double?
+    let currentTime: Double?
+    let posterURL: String?
 }
 
 struct UpdateSpatializedDynamic3DElementProperties: SpatializedElementProperties {