+++ packages/visionOS/web-spatial/WebMsgCommand.swift
@@ -25,6 +25,8 @@ enum SpatialWebMsgType: String, Encodable {
     case spatialmagnify
     case spatialmagnifyend
 
+    case animationstatechange
+
     case objectdestroy
 }
 
@@ -112,6 +114,19 @@ struct ModelLoadSuccess: Encodable {
 
 struct ModelLoadFailure: Encodable {
     let type: SpatialWebMsgType = .modelloadfailed
+}
+
+struct AnimationStateChangeDetail: Encodable {
+    let paused: Bool
+    let duration: Double
+    let currentTime: Double
+    /// Unix epoch time in milliseconds
+    let timestamp: Double
+}
+
+struct AnimationStateChangeEvent: Encodable {
+    let type: SpatialWebMsgType = .animationstatechange
+    let detail: AnimationStateChangeDetail
 }
 
 struct SpatialObjectDestroiedEvent: Encodable {